+++ packages/components/containers/error/GenericError.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 import { c } from 'ttag';
 
-import errorImg from '@proton/styles/assets/img/errors/generic-error.svg';
+import errorImg from '@proton/styles/assets/img/errors/error-generic.svg';
 
 import IllustrationPlaceholder from '../illustration/IllustrationPlaceholder';
 import InlineLinkButton from '../../components/button/InlineLinkButton';