+++ packages/pack/webpack/css.loader.js
@@ -8,7 +8,7 @@ const SASS_VARIABLES_FILEPATH = path.resolve('./src/app/variables.scss');
 const SASS_VARIABLES = fs.existsSync(SASS_VARIABLES_FILEPATH) ? fs.readFileSync(SASS_VARIABLES_FILEPATH) : '';
 // Set up the variables to the design system so that files are resolved properly.
 const PREPEND_SASS = `
-$path-images: "~@proton/styles/assets/img/shared/";
+$path-images: "~@proton/styles/assets/img/illustrations/";
 ${SASS_VARIABLES}
 `;